add registry unit
