+++ spi-samples-configcenter/spi-samples-configcenter-etcd/src/test/java/configcenter/DemoServiceIT.java
@@ -18,7 +18,7 @@
 package configcenter;
 
 
-import org.apache.dubbo.spi.samples.configcenter.api.DemoService;
+import org.apache.dubbo.spi.samples.registry.api.DemoService;
 
 import org.junit.Assert;
 import org.junit.Test;
