+++ spi-samples-registry/spi-samples-registry-etcd/case-configuration.yml
@@ -1,0 +1,63 @@
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+timeout: 600
+
+services:
+  dubbo-spi-samples-registry-etcd:
+    type: app
+    basedir: .
+    mainClass: org.apache.dubbo.spi.samples.configcenter.EtcdProvider
+    systemProps:
+      - etcd.address=etcd
+      - etcd.port=2379
+    waitPortsBeforeRun:
+      - etcd:2379
+    waitTimeout: 100
+    checkPorts:
+      - 20880
+    checkLog: "dubbo service started"
+    depends_on:
+      - etcd
+
+  dubbo-spi-samples-registry-etcd-test:
+    type: test
+    basedir: .
+    tests:
+      - "**/*IT.class"
+    systemProps:
+      - dubbo.address=dubbo-spi-samples-registry-etcd
+      - dubbo.port=20880
+      - etcd.address=etcd
+      - etcd.port=2379
+    waitPortsBeforeRun:
+      - dubbo-spi-samples-registry-etcd:2181
+      - dubbo-spi-samples-registry-etcd:20880
+      - etcd:2379
+    waitTimeout: 100
+    depends_on:
+      - etcd
+      - dubbo-spi-samples-registry-etcd
+
+  etcd:
+    image: bitnami/etcd:3.5.11
+    environment:
+      - ETCD_ENABLE_V2=true
+      - ALLOW_NONE_AUTHENTICATION=yes
+      - ETCD_ADVERTISE_CLIENT_URLS=http://etcd:2379
+      - ETCD_LISTEN_CLIENT_URLS=http://0.0.0.0:2379
+    expose:
+      - 2379
