+++ spi-samples-registry/spi-samples-registry-etcd/src/main/resources/spring/configcenter-consumer.xml
@@ -25,11 +25,11 @@
        http://dubbo.apache.org/schema/dubbo http://dubbo.apache.org/schema/dubbo/dubbo.xsd http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context.xsd">
     <context:property-placeholder/>
 
-    <dubbo:application name="dubbo-spi-configcenter-etcd-consumer"/>
+    <dubbo:application name="dubbo-spi-registry-etcd-consumer"/>
 
-    <dubbo:registry address="consul://${consul.address:127.0.0.1}:${consul.port:8500}"/>
+    <dubbo:registry address="etcd3://${etcd.address:127.0.0.1}:${etcd.port:2379}"/>
 
-    <dubbo:config-center highest-priority="false" address="consul://${consul.address:127.0.0.1}:${consul.port:8500}"/>
+    <dubbo:config-center highest-priority="false" address="etcd3://${etcd.address:localhost}:${etcd.port:2379}"/>
 
     <dubbo:reference id="demoService" interface="org.apache.dubbo.spi.samples.registry.api.DemoService"/>
